% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpd.R
\name{gpd}
\alias{gpd}
\alias{dgpd}
\alias{pgpd}
\alias{qgpd}
\alias{rgpd}
\title{Generalised (Type II) pareto distribution}
\usage{
dgpd(x, a, b, log = FALSE)

pgpd(x, a, b)

qgpd(q, a, b)

rgpd(n, a, b)
}
\arguments{
\item{x}{vector of samples}

\item{a, b}{the shape and scale parameters of the GPD}

\item{log}{Logical if TRUE returns log of density}

\item{q}{vector of quantiles}

\item{n}{number of samples to draw}
}
\description{
Functions for evaluating the density, cumulative density and quantilies of the GPD along with random number generation.
}
\details{
For \eqn{b\ge0} with \eqn{x\ge0} when \eqn{c\le0} or
\eqn{0\lex\le\frac{b/a}} when \eqn{c\ge0}, the density function is given by
\deqn{ f(x)=\frac{1}{b}(1-\frac{a}{b}x)^{\frac{1}{a}-1} }
For \eqn{a} not equal to \eqn{0} the corresponding distribution function is
\deqn{ F(x) = 1-(1-\frac{a}{b}x)^{\frac{1}{a}} }
while for \eqn{a=0}
\deqn{F(x) = 1-\exp(1-\frac{x}{b})}
}
\examples{
x <- rgpd(3,c(-1,0,1),2)
p <- pgpd(x,c(-1,0,1),2)
xx <- qgpd(p,c(-1,0,1),2)
stopifnot( all(x==xx) )
d <- dgpd(x,c(-1,0,1),2)

}
