% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep_objects.R
\name{keep_objects}
\alias{keep_objects}
\title{Keep selected objects}
\usage{
keep_objects(x)
}
\arguments{
\item{x}{Character vector of object names}
}
\value{
NA
}
\description{
Keep selected objects in the environment, remove everything else.
}
\examples{
o1 <- "Object 1"; #' o2 <- "Object 2"; #' o3 <- "Object 3" 
keep_objects(c("o1","o2"))
keep_objects("o1")

}
