% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksgpd.R
\name{ksgpd}
\alias{ksgpd}
\alias{plot_ksgpd}
\alias{dksgpd}
\alias{pksgpd}
\alias{qksgpd}
\title{Empirical distribution with kernel smoothing for bulk and Generalised (Type II) pareto distribution upper tail}
\usage{
ksgpd(x, numBreak = 1000, skipBreak = 10, mina = 1e-10, keepEstimate)

plot_ksgpd(
  dst,
  xlb = expression(paste(paste("Discharge [", m^3), "/s]")),
  ylb = "Profile Likelihood",
  ttl = ""
)

dksgpd(x, dst, log = FALSE)

pksgpd(x, dst)

qksgpd(q, dst)
}
\arguments{
\item{x}{vector of samples}

\item{numBreak}{Maximum number of breaks to consider. The values are the numBreak highest values in the sample}

\item{skipBreak}{Miss the skipBreak largest values out when testing break points}

\item{mina}{Minimum value of the GPD shape parameter to try (a<=0 gives infinite tails to the distribution)}

\item{keepEstimate}{Flag to indicate is estimation information should be kept}

\item{dst}{A ksgpd object}

\item{xlb}{x axes label for plot [expression(paste(paste("Discharge [",m^3),"/s]"))]

[",m^3),"/s]: R:\%22,m\%5E3),\%22/s}

\item{ylb}{y axes label for plot \link{'Profile Likelihood'}}

\item{ttl}{Title of the plot \link{""}}

\item{log}{Logical if TRUE returns log density \link{FALSE}}

\item{q}{vector of quantiles}
}
\value{
In the case of ksgpd a list object containing
brk, Pbrk - the selecte break value and the distribution function value at that point
gpd - a list with the shape (a) and scale (b) of the GPD tail of the distribution
ks - a list with the centers (cntr) and bandwidth (bw) of the kernal smooth approximation
estimate - a list of tried break points (brks) and the corresponding log likelihoos (llkl)
}
\description{
Functions for evaluating the density, cumulative density and quantilies of the distribution.
}
