% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_str.R
\name{tbl_str}
\alias{tbl_str}
\title{Data frame descriptives}
\usage{
tbl_str(x, ordered = FALSE, n_examples = 3, truncate_examples_at = 50)
}
\arguments{
\item{x}{Data frame}

\item{ordered}{Set to TRUE if the table is to be ordered alphabetically by the
variable names}

\item{n_examples}{Number of examples of the values of each variable to show in
the table; the default value is 3 examples.}

\item{truncate_examples_at}{Number of characters after which the examples of each
variable are cut off. This is to prevent examples that are long strings to mess
us displaying the table; by default, all examples are truncated at 50 characters.}
}
\value{
A data frame containing basic descriptive statistics
}
\description{
A table of basic descriptive statistics for a given data frame.
}
\details{
\code{tbl_str} returns a table of basic descriptive statistics for a given
data frame. In terms of R objects, the returned table is itself a data frame and can
be manipulated as such. Most of the table's columns are self-explanatory, 
\code{n} is the observation count, i.e., the number of non-missing observations,
\code{n_na} is the number of missing observations, and \code{n_unique} is the number
}
\examples{
# Generate a data frame with different types of variables; also introduce a handful
of missing values

model <- rownames(mtcars)
mpg   <- as.double(mtcars$mpg)
cyl   <- as.integer(mtcars$cyl)
am    <- as.logical(mtcars$am)
gear  <- as.ordered(mtcars$gear)

df <- cbind.data.frame(model, mpg, cyl, am, gear, row.names = NULL,
      stringsAsFactors = FALSE);
      df[1:5, 2] <- NA; df[11:20, 3] <- NA

tbl_str(df, ordered = TRUE, n_examples = 3)

}
