% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_sum_dates.R
\name{tbl_sum_dates}
\alias{tbl_sum_dates}
\title{Summary table for class date variables}
\usage{
tbl_sum_dates(x, sorted = FALSE)
}
\arguments{
\item{x}{Data frame containing only date variables}

\item{sorted}{TRUE if the table should be sorted by the variable names, FALSE if
otherwise (the default).}
}
\value{
A data frame consisting of statistics typically used to describe date
variables.
}
\description{
Table of selected summary statistics for date variables
}
\details{
The argument must be a data frame (if it's an atomic vector, coerce it into
data frame first), and it must contain only date variables. If non-date variables are
present, an error message 'All variables must be class date' will show.
}
\examples{
df <- data.frame( x1 = c(1:9)*0.1,
                  x2 = as.integer(c(1:9)*5),
                  x3 = c(1:9)*10,
                  x4 = seq(from = as.Date("1970-01-01"), by = "day", length.out = 9),
                  x5 = letters[1:9],
                  x6 = as.factor(sample(c("f1", "f2", "f3"), 9, replace = TRUE)),
                  x7 = seq(from = as.Date("2018-05-15"), by = "5 days", length.out = 9),
                  stringsAsFactors = FALSE )

df[2:3, 1] <- NA; df[3:5, 2] <- NA; df[4:5, 3] <- NA; df[c(3,5), 4] <- NA

date_vars <- unlist(lapply(df,class)) == "Date" # Identify date variables
df_dates <- df[, date_vars] # Data frame of only date variables

tbl_sum_dates(df) # Error is returned because of non-date variables
tbl_sum_dates(df_dates)
tbl_sum_dates(df_dates, sorted = TRUE)

}
