% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_to_execute.R
\name{time_to_execute}
\alias{time_to_execute}
\title{Duration of executing an R code}
\usage{
time_to_execute(..., which_time = "elapsed")
}
\arguments{
\item{...}{All R code to be executed}
}
\value{
Duration of code execution in seconds
}
\description{
Time it takes to execute a given code.
}
\details{
Returns how many seconds it took to exectue a given
line/block of code. Note that if the code extends over multiple lines,
it has to be enclosed in curly brackets.
}
\examples{
x <- matrix(runif(1e7), ncol = 1e5)

time_to_execute( unlist(lapply(list(x), colMeans)) )

# Curly brackets for multiple lines of code:
time_to_execute( which_time = "elapsed",
                 { out_loop <- c()
                  for (i in seq_len(ncol(x))) { out_loop[[i]] <- mean(x[ , i]) }
                 } )

}
